+++ ml/train/preprocessing/annotation_generator.py
@@ -11,7 +11,7 @@ from botocore.exceptions import ClientError
 from keras_retinanet.preprocessing.csv_generator import Generator
 from keras_retinanet.utils.image import read_image_bgr
 
-import config
+import config.config
 from utils.query import pd_query, cursor
