+++ ml/utils/evaluate.py
@@ -1,7 +1,7 @@
-from model_scoring import f1_evaluation
+from train.evaluation.model_scoring import f1_evaluation
 from keras_retinanet.models import convert_model
 
-import config
+import config.config
 
 
 def evaluate_class_thresholds(model, generator):
@@ -13,7 +13,8 @@ def evaluate_class_thresholds(model, generator):
     # Initializing model for eval
     model = convert_model(model)
 
-    best_f1, best_thresh = f1_evaluation(generator, model, save_path=config.TEST_EXAMPLES)
+    best_f1, best_thresh = f1_evaluation(
+        generator, model, save_path=config.TEST_EXAMPLES)
 
     total_f1 = 0
     for concept, f1 in best_f1.items():
@@ -30,4 +31,5 @@ if __name__ == '__main__':
     concepts = [1629, 1210, 236, 383, 1133]
     model_path = 'current_weights.h5'
 
-    evaluate_class_thresholds(concepts, model_path, min_examples, download_data=False)
+    evaluate_class_thresholds(concepts, model_path,
+                              min_examples, download_data=False)
